Updated Code

+++ app/src/main/java/com/clebs/celerity/models/requests/AddOnRouteInfoRequest.kt
@@ -4,11 +4,11 @@ data class AddOnRouteInfoRequest(
     val RtAddMode: String,
     val RtComment: String,
     val RtDwId: Int,
-    val RtFinishMileage: Int,
+    val RtFinishMileage: Long,
     val RtTypeId: Int,
     val RtLocationId: Int,
     val RtName: String,
-    val RtNoOfParcelsDelivered: Int,
+    val RtNoOfParcelsDelivered: Long,
     val RtNoParcelsbroughtback: Int,
     val RtUsrId: Int,
     val VehicleId: Int
